Restructured the mod version folders
Now instead of having 3 independently installable versions of the mod:
- A version with no spoilery content at all
- A version with the the spoiler-free content and also base-game spoilers
- A version with content from the previous 2 versions, plus DLC spoilers

Now there are 4 "modules":
- A module with no spoilery content at all
- A module with the slightly spoilery nomai creatures and civ
- A module with the spoilery content from the base game, except the nomai
- A module with the spoilery content from the Echoes of the Eye DLC

This comes at the cost of making it slightly more cumbersome to install
the full version of the mod, but makes development less annoying/involve
less duplication of changes, and hopefully will make publishing releases
for the mod easier in the future, assuming Tarn implements an ingame
"mod config" menu of some kind.

Obviously also updated the readme to reflect this change, though there
are minor, unrelated changes to the readme as well.